+++ src/Clapeyron.jl
@@ -9,7 +9,6 @@ const PackedVofV = PackedVectorsOfVectors.PackedVectorOfVectors
 using Roots: Roots
 
 using Scratch 
-using Unitful
 import LogExpFunctions
 using FillArrays: FillArrays
 import BlackBoxOptim
@@ -18,6 +17,7 @@ using NLSolvers
 using NLSolvers: NEqOptions
 using DiffResults, ForwardDiff
 using Downloads #for bibtex
+
 #compatibility and raw julia utilities
 include("utils/core_utils.jl")
 
@@ -91,6 +91,10 @@ include("utils/acceleration_ss.jl")
 #Clapeyron methods (AD, property solvers, etc)
 include("methods/methods.jl")
 
+#Unitful support, transition from dependency to ext
+if !isdefined(Base,:get_extension)
+    include("../ext/ClapeyronUnitfulExt.jl")
+end
 #=
 the dependency chain is the following:
 base --> database(params)  -|-> split_model --> methods -|-> models