+++ src/methods/methods.jl
@@ -218,6 +218,7 @@ function gradient_type(V,T,z::FractionVector)
     return Vector{μ}
 end
 
+
 """
     init_preferred_method(method,model,kwargs)
 
@@ -235,5 +236,3 @@ include("property_solvers/property_solvers.jl")
 include("tpd.jl")
 include("stability.jl")
 include("pT.jl")
-include("unitful_base.jl")
-include("unitful_methods.jl")