+++ NEWS.md
@@ -1,7 +1,10 @@
 # v0.5.6
 
 ## New Features
-- `Clapeyron.diagvalues` now accepts `x::Number` (returning the same number)
+- New EoS: Polar PCSAFT with Quadrupolar interactions (`QPPCSAFT`)
+- `Unitful.jl` support has been moved into an extension.
+- `MultiComponentFlash.jl` support via extension: you can pass `model::EoSModel` to `MultiComponentFlash.flash_2ph`. there is also a new `MCFlashJL` method that calls `flash_2ph` using the `Clapeyron.tp_flash` interface.
+- custom types can be passed to the `userlocations` keyword argument, defining `Clapeyron.can_nt(::datatype) = true` and `Clapeyron.to_nt(x::datatype)::Union{AbstractDict,NamedTuple}`
 
 ## Bug Fixes
-- more flexible sites parser (#214). Before, some site names where hardcoded.
+- bug in `pharmaPCSAFT` mixing rules